Use templates; change files structure; fix XSS
